# Conflicts:
#	app/src/main/AndroidManifest.xml
#	app/src/main/java/kr/dori/android/own_cast/forApiData/AuthRetrofitInterface.kt
#	app/src/main/java/kr/dori/android/own_cast/playlist/CastFragment.kt
#	app/src/main/java/kr/dori/android/own_cast/playlist/CategoryFragment.kt
#	app/src/main/java/kr/dori/android/own_cast/playlist/PlaylistCategoryAdapter.kt
#	app/src/main/java/kr/dori/android/own_cast/playlist/PlaylistFragment.kt

+++ app/src/main/java/kr/dori/android/own_cast/forApiData/AuthRetrofitInterface.kt
@@ -113,7 +113,9 @@ interface Bookmark {
             Response<AuthResponse<Long>>
 
     @GET("/api/study/{playlistId}")
-    suspend fun getBookmark(@Path("playlistId") playlistId: Int):
+
+    suspend fun getBookmark(@Path("playlistId")playlistId: Long):
+
             Response<AuthResponse<List<GetBookmark>>>
 
     @GET("/api/study/savedcast")
